Update on main code
Updated code to fit suggestions people made on the pilot study. Added goodbye page to instruct participants to close the tab, fixed some wording issues and typos, added an option for extended tutorial on how to access DIALANG, and organized where experimental data is displayed on the exported data file.